+++ internal-packages/run-engine/src/engine/systems/batchSystem.ts
@@ -3,29 +3,21 @@ import { Logger } from "@trigger.dev/core/logger";
 import { PrismaClient } from "@trigger.dev/database";
 import { isFinalRunStatus } from "../statuses.js";
 import { EngineWorker } from "../types.js";
+import { SystemResources } from "./systems.js";
 
 export type BatchSystemOptions = {
-  prisma: PrismaClient;
-  logger: Logger;
-  tracer: Tracer;
-  worker: EngineWorker;
+  resources: SystemResources;
 };
 
 export class BatchSystem {
-  private readonly prisma: PrismaClient;
-  private readonly logger: Logger;
-  private readonly tracer: Tracer;
-  private readonly worker: EngineWorker;
+  private readonly $: SystemResources;
 
   constructor(private readonly options: BatchSystemOptions) {
-    this.prisma = options.prisma;
-    this.logger = options.logger;
-    this.tracer = options.tracer;
-    this.worker = options.worker;
+    this.$ = options.resources;
   }
 
   public async scheduleCompleteBatch({ batchId }: { batchId: string }): Promise<void> {
-    await this.worker.enqueue({
+    await this.$.worker.enqueue({
       //this will debounce the call
       id: `tryCompleteBatch:${batchId}`,
       job: "tryCompleteBatch",
@@ -44,8 +36,8 @@ export class BatchSystem {
    * This isn't used operationally, but it's used for the Batches dashboard page.
    */
   async #tryCompleteBatch({ batchId }: { batchId: string }) {
-    return startSpan(this.tracer, "#tryCompleteBatch", async (span) => {
-      const batch = await this.prisma.batchTaskRun.findUnique({
+    return startSpan(this.$.tracer, "#tryCompleteBatch", async (span) => {
+      const batch = await this.$.prisma.batchTaskRun.findUnique({
         select: {
           status: true,
           runtimeEnvironmentId: true,
@@ -56,16 +48,16 @@ export class BatchSystem {
       });
 
       if (!batch) {
-        this.logger.error("#tryCompleteBatch batch doesn't exist", { batchId });
+        this.$.logger.error("#tryCompleteBatch batch doesn't exist", { batchId });
         return;
       }
 
       if (batch.status === "COMPLETED") {
-        this.logger.debug("#tryCompleteBatch: Batch already completed", { batchId });
+        this.$.logger.debug("#tryCompleteBatch: Batch already completed", { batchId });
         return;
       }
 
-      const runs = await this.prisma.taskRun.findMany({
+      const runs = await this.$.prisma.taskRun.findMany({
         select: {
           id: true,
           status: true,
@@ -77,8 +69,8 @@ export class BatchSystem {
       });
 
       if (runs.every((r) => isFinalRunStatus(r.status))) {
-        this.logger.debug("#tryCompleteBatch: All runs are completed", { batchId });
-        await this.prisma.batchTaskRun.update({
+        this.$.logger.debug("#tryCompleteBatch: All runs are completed", { batchId });
+        await this.$.prisma.batchTaskRun.update({
           where: {
             id: batchId,
           },
@@ -87,7 +79,7 @@ export class BatchSystem {
           },
         });
       } else {
-        this.logger.debug("#tryCompleteBatch: Not all runs are completed", { batchId });
+        this.$.logger.debug("#tryCompleteBatch: Not all runs are completed", { batchId });
       }
     });
   }
