+++ internal-packages/run-engine/src/engine/systems/systems.ts
@@ -1,0 +1,23 @@
+import { Tracer } from "@internal/tracing";
+import { Logger } from "@trigger.dev/core/logger";
+import { PrismaClient } from "@trigger.dev/database";
+import { RunQueue } from "../../run-queue/index.js";
+import { EventBus } from "../eventBus.js";
+import { RunLocker } from "../locking.js";
+import { EngineWorker } from "../types.js";
+import { ReleaseConcurrencyTokenBucketQueue } from "../releaseConcurrencyTokenBucketQueue.js";
+
+export type SystemResources = {
+  prisma: PrismaClient;
+  worker: EngineWorker;
+  eventBus: EventBus;
+  logger: Logger;
+  tracer: Tracer;
+  runLock: RunLocker;
+  runQueue: RunQueue;
+  releaseConcurrencyQueue: ReleaseConcurrencyTokenBucketQueue<{
+    orgId: string;
+    projectId: string;
+    envId: string;
+  }>;
+};
